+++ scripts/rollback/rollback-3.13.4.sh
@@ -1,0 +1,88 @@
+#!/bin/bash
+# timestamp: 1701574047001
+
+npm dist-tag add @midwayjs/axios@3.12.9 latest
+npm dist-tag add @midwayjs/bull@3.12.8 latest
+npm dist-tag add @midwayjs/bull-board@3.12.8 latest
+npm dist-tag add @midwayjs/captcha@3.12.8 latest
+npm dist-tag add @midwayjs/casbin@3.12.8 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.12.8 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.12.8 latest
+npm dist-tag add @midwayjs/code-dye@3.12.8 latest
+npm dist-tag add @midwayjs/consul@3.12.9 latest
+npm dist-tag add @midwayjs/cron@3.12.8 latest
+npm dist-tag add @midwayjs/cross-domain@3.12.8 latest
+npm dist-tag add @midwayjs/grpc@3.12.8 latest
+npm dist-tag add @midwayjs/http-proxy@3.12.8 latest
+npm dist-tag add @midwayjs/i18n@3.12.8 latest
+npm dist-tag add @midwayjs/info@3.12.8 latest
+npm dist-tag add @midwayjs/jwt@3.12.8 latest
+npm dist-tag add @midwayjs/mongoose@3.12.9 latest
+npm dist-tag add @midwayjs/otel@3.12.8 latest
+npm dist-tag add @midwayjs/passport@3.12.8 latest
+npm dist-tag add @midwayjs/process-agent@3.12.8 latest
+npm dist-tag add @midwayjs/prometheus@3.12.8 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.12.8 latest
+npm dist-tag add @midwayjs/rabbitmq@3.12.8 latest
+npm dist-tag add @midwayjs/security@3.12.8 latest
+npm dist-tag add @midwayjs/sequelize@3.12.8 latest
+npm dist-tag add @midwayjs/socketio@3.12.8 latest
+npm dist-tag add @midwayjs/static-file@3.12.8 latest
+npm dist-tag add @midwayjs/swagger@3.12.9 latest
+npm dist-tag add @midwayjs/tags@3.12.8 latest
+npm dist-tag add @midwayjs/typegoose@3.12.9 latest
+npm dist-tag add @midwayjs/typeorm@3.12.8 latest
+npm dist-tag add @midwayjs/upload@3.12.8 latest
+npm dist-tag add @midwayjs/validate@3.12.8 latest
+npm dist-tag add @midwayjs/version@3.12.9 latest
+npm dist-tag add @midwayjs/view@3.12.8 latest
+npm dist-tag add @midwayjs/view-ejs@3.12.8 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.12.8 latest
+npm dist-tag add @midwayjs/web@3.12.8 latest
+npm dist-tag add @midwayjs/express@3.12.8 latest
+npm dist-tag add @midwayjs/koa@3.12.8 latest
+npm dist-tag add @midwayjs/ws@3.12.8 latest
+
+# Changes:
+
+#  - @midwayjs/axios: 3.12.9 => 3.13.4
+#  - @midwayjs/bull: 3.12.8 => 3.13.4
+#  - @midwayjs/bull-board: 3.12.8 => 3.13.4
+#  - @midwayjs/captcha: 3.12.8 => 3.13.4
+#  - @midwayjs/casbin: 3.12.8 => 3.13.4
+#  - @midwayjs/casbin-redis-adapter: 3.12.8 => 3.13.4
+#  - @midwayjs/casbin-typeorm-adapter: 3.12.8 => 3.13.4
+#  - @midwayjs/code-dye: 3.12.8 => 3.13.4
+#  - @midwayjs/consul: 3.12.9 => 3.13.4
+#  - @midwayjs/cron: 3.12.8 => 3.13.4
+#  - @midwayjs/cross-domain: 3.12.8 => 3.13.4
+#  - @midwayjs/grpc: 3.12.8 => 3.13.4
+#  - @midwayjs/http-proxy: 3.12.8 => 3.13.4
+#  - @midwayjs/i18n: 3.12.8 => 3.13.4
+#  - @midwayjs/info: 3.12.8 => 3.13.4
+#  - @midwayjs/jwt: 3.12.8 => 3.13.4
+#  - @midwayjs/mongoose: 3.12.9 => 3.13.4
+#  - @midwayjs/otel: 3.12.8 => 3.13.4
+#  - @midwayjs/passport: 3.12.8 => 3.13.4
+#  - @midwayjs/process-agent: 3.12.8 => 3.13.4
+#  - @midwayjs/prometheus: 3.12.8 => 3.13.4
+#  - @midwayjs/prometheus-socket-io: 3.12.8 => 3.13.4
+#  - @midwayjs/rabbitmq: 3.12.8 => 3.13.4
+#  - @midwayjs/security: 3.12.8 => 3.13.4
+#  - @midwayjs/sequelize: 3.12.8 => 3.13.4
+#  - @midwayjs/socketio: 3.12.8 => 3.13.4
+#  - @midwayjs/static-file: 3.12.8 => 3.13.4
+#  - @midwayjs/swagger: 3.12.9 => 3.13.4
+#  - @midwayjs/tags: 3.12.8 => 3.13.4
+#  - @midwayjs/typegoose: 3.12.9 => 3.13.4
+#  - @midwayjs/typeorm: 3.12.8 => 3.13.4
+#  - @midwayjs/upload: 3.12.8 => 3.13.4
+#  - @midwayjs/validate: 3.12.8 => 3.13.4
+#  - @midwayjs/version: 3.12.9 => 3.13.4
+#  - @midwayjs/view: 3.12.8 => 3.13.4
+#  - @midwayjs/view-ejs: 3.12.8 => 3.13.4
+#  - @midwayjs/view-nunjucks: 3.12.8 => 3.13.4
+#  - @midwayjs/web: 3.12.8 => 3.13.4
+#  - @midwayjs/express: 3.12.8 => 3.13.4
+#  - @midwayjs/koa: 3.12.8 => 3.13.4
+#  - @midwayjs/ws: 3.12.8 => 3.13.4