+++ scripts/rollback/rollback-3.13.6.sh
@@ -1,0 +1,126 @@
+#!/bin/bash
+# timestamp: 1702570777903
+
+npm dist-tag add @midwayjs/fc-starter@3.13.5 latest
+npm dist-tag add @midwayjs/serverless-http-parser@3.12.5 latest
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.13.5 latest
+npm dist-tag add @midwayjs/axios@3.13.5 latest
+npm dist-tag add @midwayjs/bootstrap@3.13.5 latest
+npm dist-tag add @midwayjs/bull@3.13.5 latest
+npm dist-tag add @midwayjs/bull-board@3.13.5 latest
+npm dist-tag add @midwayjs/cache@3.13.5 latest
+npm dist-tag add @midwayjs/captcha@3.13.5 latest
+npm dist-tag add @midwayjs/casbin@3.13.5 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.13.5 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.13.5 latest
+npm dist-tag add @midwayjs/code-dye@3.13.5 latest
+npm dist-tag add @midwayjs/consul@3.13.5 latest
+npm dist-tag add @midwayjs/core@3.13.5 latest
+npm dist-tag add @midwayjs/cos@3.13.5 latest
+npm dist-tag add @midwayjs/cron@3.13.5 latest
+npm dist-tag add @midwayjs/cross-domain@3.13.5 latest
+npm dist-tag add @midwayjs/decorator@3.13.5 latest
+npm dist-tag add @midwayjs/etcd@3.13.5 latest
+npm dist-tag add @midwayjs/express-session@3.13.5 latest
+npm dist-tag add @midwayjs/faas@3.13.5 latest
+npm dist-tag add @midwayjs/grpc@3.13.5 latest
+npm dist-tag add @midwayjs/http-proxy@3.13.5 latest
+npm dist-tag add @midwayjs/i18n@3.13.5 latest
+npm dist-tag add @midwayjs/info@3.13.5 latest
+npm dist-tag add @midwayjs/jwt@3.13.5 latest
+npm dist-tag add @midwayjs/kafka@3.13.5 latest
+npm dist-tag add @midwayjs/leoric@3.13.5 latest
+npm dist-tag add @midwayjs/mikro@3.13.5 latest
+npm dist-tag add @midwayjs/mock@3.13.5 latest
+npm dist-tag add @midwayjs/mongoose@3.13.5 latest
+npm dist-tag add @midwayjs/oss@3.13.5 latest
+npm dist-tag add @midwayjs/otel@3.13.5 latest
+npm dist-tag add @midwayjs/passport@3.13.5 latest
+npm dist-tag add @midwayjs/process-agent@3.13.5 latest
+npm dist-tag add @midwayjs/prometheus@3.13.5 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.13.5 latest
+npm dist-tag add @midwayjs/rabbitmq@3.13.5 latest
+npm dist-tag add @midwayjs/redis@3.13.5 latest
+npm dist-tag add @midwayjs/security@3.13.5 latest
+npm dist-tag add @midwayjs/sequelize@3.13.5 latest
+npm dist-tag add @midwayjs/session@3.13.5 latest
+npm dist-tag add @midwayjs/socketio@3.13.5 latest
+npm dist-tag add @midwayjs/static-file@3.13.5 latest
+npm dist-tag add @midwayjs/swagger@3.13.5 latest
+npm dist-tag add @midwayjs/tablestore@3.13.5 latest
+npm dist-tag add @midwayjs/tags@3.13.5 latest
+npm dist-tag add @midwayjs/typegoose@3.13.5 latest
+npm dist-tag add @midwayjs/typeorm@3.13.5 latest
+npm dist-tag add @midwayjs/upload@3.13.5 latest
+npm dist-tag add @midwayjs/validate@3.13.5 latest
+npm dist-tag add @midwayjs/version@3.13.5 latest
+npm dist-tag add @midwayjs/view@3.13.5 latest
+npm dist-tag add @midwayjs/view-ejs@3.13.5 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.13.5 latest
+npm dist-tag add @midwayjs/web@3.13.5 latest
+npm dist-tag add @midwayjs/express@3.13.5 latest
+npm dist-tag add @midwayjs/koa@3.13.5 latest
+npm dist-tag add @midwayjs/ws@3.13.5 latest
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.13.5 => 3.13.6
+#  - @midwayjs/serverless-http-parser: 3.12.5 => 3.13.6
+#  - @midwayjs/async-hooks-context-manager: 3.13.5 => 3.13.6
+#  - @midwayjs/axios: 3.13.5 => 3.13.6
+#  - @midwayjs/bootstrap: 3.13.5 => 3.13.6
+#  - @midwayjs/bull: 3.13.5 => 3.13.6
+#  - @midwayjs/bull-board: 3.13.5 => 3.13.6
+#  - @midwayjs/cache: 3.13.5 => 3.13.6
+#  - @midwayjs/captcha: 3.13.5 => 3.13.6
+#  - @midwayjs/casbin: 3.13.5 => 3.13.6
+#  - @midwayjs/casbin-redis-adapter: 3.13.5 => 3.13.6
+#  - @midwayjs/casbin-typeorm-adapter: 3.13.5 => 3.13.6
+#  - @midwayjs/code-dye: 3.13.5 => 3.13.6
+#  - @midwayjs/consul: 3.13.5 => 3.13.6
+#  - @midwayjs/core: 3.13.5 => 3.13.6
+#  - @midwayjs/cos: 3.13.5 => 3.13.6
+#  - @midwayjs/cron: 3.13.5 => 3.13.6
+#  - @midwayjs/cross-domain: 3.13.5 => 3.13.6
+#  - @midwayjs/decorator: 3.13.5 => 3.13.6
+#  - @midwayjs/etcd: 3.13.5 => 3.13.6
+#  - @midwayjs/express-session: 3.13.5 => 3.13.6
+#  - @midwayjs/faas: 3.13.5 => 3.13.6
+#  - @midwayjs/grpc: 3.13.5 => 3.13.6
+#  - @midwayjs/http-proxy: 3.13.5 => 3.13.6
+#  - @midwayjs/i18n: 3.13.5 => 3.13.6
+#  - @midwayjs/info: 3.13.5 => 3.13.6
+#  - @midwayjs/jwt: 3.13.5 => 3.13.6
+#  - @midwayjs/kafka: 3.13.5 => 3.13.6
+#  - @midwayjs/leoric: 3.13.5 => 3.13.6
+#  - @midwayjs/mikro: 3.13.5 => 3.13.6
+#  - @midwayjs/mock: 3.13.5 => 3.13.6
+#  - @midwayjs/mongoose: 3.13.5 => 3.13.6
+#  - @midwayjs/oss: 3.13.5 => 3.13.6
+#  - @midwayjs/otel: 3.13.5 => 3.13.6
+#  - @midwayjs/passport: 3.13.5 => 3.13.6
+#  - @midwayjs/process-agent: 3.13.5 => 3.13.6
+#  - @midwayjs/prometheus: 3.13.5 => 3.13.6
+#  - @midwayjs/prometheus-socket-io: 3.13.5 => 3.13.6
+#  - @midwayjs/rabbitmq: 3.13.5 => 3.13.6
+#  - @midwayjs/redis: 3.13.5 => 3.13.6
+#  - @midwayjs/security: 3.13.5 => 3.13.6
+#  - @midwayjs/sequelize: 3.13.5 => 3.13.6
+#  - @midwayjs/session: 3.13.5 => 3.13.6
+#  - @midwayjs/socketio: 3.13.5 => 3.13.6
+#  - @midwayjs/static-file: 3.13.5 => 3.13.6
+#  - @midwayjs/swagger: 3.13.5 => 3.13.6
+#  - @midwayjs/tablestore: 3.13.5 => 3.13.6
+#  - @midwayjs/tags: 3.13.5 => 3.13.6
+#  - @midwayjs/typegoose: 3.13.5 => 3.13.6
+#  - @midwayjs/typeorm: 3.13.5 => 3.13.6
+#  - @midwayjs/upload: 3.13.5 => 3.13.6
+#  - @midwayjs/validate: 3.13.5 => 3.13.6
+#  - @midwayjs/version: 3.13.5 => 3.13.6
+#  - @midwayjs/view: 3.13.5 => 3.13.6
+#  - @midwayjs/view-ejs: 3.13.5 => 3.13.6
+#  - @midwayjs/view-nunjucks: 3.13.5 => 3.13.6
+#  - @midwayjs/web: 3.13.5 => 3.13.6
+#  - @midwayjs/express: 3.13.5 => 3.13.6
+#  - @midwayjs/koa: 3.13.5 => 3.13.6
+#  - @midwayjs/ws: 3.13.5 => 3.13.6